Let a linkshell be worn
The linkshell is slot sixteen - SLOT_LINK1 - and nothing in the client could
say so. The DAT's own slot mask is sixteen bits wide, so bit sixteen does not
fit in it, and the field the renderer was given was the same width. A
linkshell therefore read as Slots 0, IsEquipment false, and the menu never
offered to put it on.

The file tells a linkshell apart by item type instead: Linkshell (513) and
Linkpearl (515) both carry type 6, where a crystal carries 8. So FfxiItem
gains IsLinkshell off that, IsWearable for either route, and the slot mask
handed to the renderer is widened to thirty-two bits with bit sixteen set on
the way through. That widening is a deliberate change to the interop
signature, and it is the smallest thing that can express a slot the file
cannot.

The equipment slots now run to Linkshell2 as the server's do, and the worn
walk with them.

What is not done: the armoury panel still draws sixteen slots, so a worn
linkshell has nowhere to show yet. Equipping one works; seeing it does not.

+++ native/moghouse_interop/include/moghouse_interop.h
@@ -243,9 +243,14 @@ MH_API void mh_viewer_set_equipment(MhViewerHandle viewer, const uint8_t* contai
 /// arrows is one icon. The pixels are RGBA, `width * height * 4` bytes, row
 /// zero at the top. The renderer copies them into its atlas and does not keep
 /// the pointer.
+///
+/// `slots` is thirty-two bits rather than the sixteen the DAT stores, because
+/// a linkshell is worn in slot sixteen and sixteen bits cannot say so. The
+/// file has no room for that bit either - it tells a linkshell apart by item
+/// type - so the client sets it and this carries it.
 MH_API void mh_viewer_push_item(MhViewerHandle viewer, uint16_t item_id,
                                 const char* name, const char* description,
-                                uint16_t type, uint16_t level, uint16_t slots,
+                                uint16_t type, uint16_t level, uint32_t slots,
                                 const uint8_t* rgba, int32_t width, int32_t height);
 
 /// What a dead player pressed in the box the renderer draws them. Matches
